Add schematics + PCB layout for a simple AMR break-out board
AMR (Audio Modem Riser) is a standard for pluggable boards with
audio codecs to connect to the AC97 interface of mainboards dating
to the late-90ies / early 2000s.

The idea of this break-out boards is to be able to interface such a
modem riser card with a microcontroller or FPGA, basically using them
as a digital interface to analog phone lines for softmodem development
as part of osmocom retronetworking.